--- src/main/java/com/hsy/likelion/LikeLionKing/domain/Likes.java
@@ -1,14 +1,0 @@
-package com.hsy.likelion.LikeLionKing.domain;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class Likes {
-    private Long id;
-    private Long memberId; // 회원 id
-    private Long postId;   //  게시물 id
-}
