+++ src/main/java/com/hsy/likelion/LikeLionKing/post/dto/PostUpdateDto.java
@@ -1,0 +1,25 @@
+package com.hsy.likelion.LikeLionKing.post.dto;
+
+import com.hsy.likelion.LikeLionKing.post.domain.Post;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class PostUpdateDto {
+    private Long id;        // 게시글 id
+    private String title;   // 게시글 제목
+    private String content; // 게시글 내용
+
+    // PostCreateDto -> Post 변환
+    public static Post toEntity(PostUpdateDto postCreateDto) {
+        Post post = Post.builder()
+                .id(postCreateDto.getId())
+                .title(postCreateDto.getTitle())
+                .content(postCreateDto.getContent())
+                .build();
+        return post;
+    }
+}
